% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemm_image_lines.R
\name{gemm_image_lines}
\alias{gemm_image_lines}
\title{Convert an image to a series of different-width lines.}
\usage{
gemm_image_lines(
  photo,
  filepath = tempfile(fileext = ".png"),
  line_color = "black",
  background_color = sample(colors(), 1)
)
}
\arguments{
\item{photo}{The path to an image to visualize, read via [magick::image_read]}

\item{filepath}{The path to save the output image to}

\item{line_color}{The color to draw the lines of the image}

\item{background_color}{The flat color to use as a background}
}
\value{
filepath
}
\description{
This function is adopted directly from  Georgios Karamanis' aRtist project
at '<https://github.com/gkaramanis/aRtist/blob/main/portraits/portraits%20line/line-portraits.R>'
}
