% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemm_random_walk.R
\name{gemm_random_walk}
\alias{gemm_random_walk}
\title{Create random walks in ggplot}
\usage{
gemm_random_walk(
  n = sample(seq(1000, 10000, 1000), 1),
  random_min = -1,
  random_max = 1,
  random_min_y = NULL,
  random_min_x = NULL,
  random_max_y = NULL,
  random_max_x = NULL,
  exponent_min = 1,
  exponent_max = 4,
  exponent_min_y = NULL,
  exponent_min_x = NULL,
  exponent_max_y = NULL,
  exponent_max_x = NULL,
  color = sample(grDevices::colors(), 1),
  fun = switch(sample(1:3, 1), sin, cos, tan),
  xfun = NULL,
  yfun = NULL,
  polar = sample(c(TRUE, FALSE), 1),
  save_args = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{Number of steps to take in the walk}

\item{random_min}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{random_max}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{random_min_y}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{random_min_x}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{random_max_y}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{random_max_x}{The minimum and maximum values that will be used to generate "seed"
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_min}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_max}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_min_y}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_min_x}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_max_y}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{exponent_max_x}{The minimum and maximum values that will be used to generate exponential
multipliers. If the x- and y- variants are NULL, they will inherit from the
untagged variants.}

\item{color}{Arguments passed to [ggplot2::geom_point].}

\item{fun}{The function to be applied to generate point values.
If the x- and y- variants are NULL, they will inherit from the
untagged variant.}

\item{xfun}{The function to be applied to generate point values.
If the x- and y- variants are NULL, they will inherit from the
untagged variant.}

\item{yfun}{The function to be applied to generate point values.
If the x- and y- variants are NULL, they will inherit from the
untagged variant.}

\item{polar}{Logical: use `coord_polar`?}

\item{save_args}{A list of arguments to pass to `ggsave`. If NULL, plots are
not saved.}

\item{seed}{The random seed to use.}
}
\description{
Create random walks in ggplot
}
