% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemm_turmite.R
\name{gemm_turmite}
\alias{gemm_turmite}
\title{Create turmites in ggplot2}
\usage{
gemm_turmite(
  n_steps = 1000,
  color = sample(grDevices::colors(), 1),
  seed = NULL
)
}
\arguments{
\item{n_steps}{The number of steps to walk the turmite through.}

\item{color}{The color to graph "success" values in (the more rare color).}

\item{seed}{The random seed to use.}
}
\description{
[Turmites](https://en.wikipedia.org/wiki/Turmite) are a type of Turing
machine, randomly walking across a 2D grid. This function simulates a
two-state relative turmite and graphs the result in ggplot2.
}
