% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemm_image_lines.R
\name{gemm_unsplash_image_lines}
\alias{gemm_unsplash_image_lines}
\title{Generate line images using Unsplash photos.}
\usage{
gemm_unsplash_image_lines(
  unsplash_token,
  filepath = tempfile(fileext = ".png"),
  line_color = "black",
  background_color = sample(colors(), 1)
)
}
\arguments{
\item{unsplash_token}{Your API access key for Unsplash}

\item{filepath}{The path to save the output image to}

\item{line_color}{The color to draw the lines of the image}

\item{background_color}{The flat color to use as a background}
}
\value{
A list containing the filepath of the output image
(as `image_output`) and the image credit (as `image_credit`)
}
\description{
Generate line images using Unsplash photos.
}
