% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmed.h.R
\name{modmed}
\alias{modmed}
\title{Moderated mediation}
\usage{
modmed(data, dep, meds, pred, xmmod, mymod, covsm, covsy,
  estMethod = "standard", bootstrap = 1000, test = TRUE,
  ci = FALSE, ciWidth = 95, pm = FALSE, paths = FALSE,
  label = FALSE, estPlot = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable}

\item{meds}{a string naming the mediator variable}

\item{pred}{a string naming the predictor variable}

\item{xmmod}{a string naming the predictor variable}

\item{mymod}{a string naming the predictor variable}

\item{covsm}{a vector with strings naming the covariates}

\item{covsy}{a vector with strings naming the covariates}

\item{estMethod}{\code{'standard'} (default), or \code{'bootstrap'}, the
estimation method to use}

\item{bootstrap}{a number between 1 and 100000 (default: 1000) specifying
the number of samples that need to been drawn in the bootstrap method}

\item{test}{\code{TRUE} (default) or \code{FALSE}, provide 'Z' and 'p'
values for the mediation estimates}

\item{ci}{\code{TRUE} or \code{FALSE} (default), provide a confidence
interval for the mediation estimates}

\item{ciWidth}{a number between 50 and 99.9 (default: 95) specifying the
confidence interval width that is used as \code{'ci'}}

\item{pm}{\code{TRUE} or \code{FALSE} (default), provide the percent
mediation effect size for the mediation estimates}

\item{paths}{\code{TRUE} or \code{FALSE} (default), provide the individual
estimates of the paths in the mediation model}

\item{label}{\code{TRUE} (default) or \code{FALSE}, provide insightful
labels for all estimates}

\item{estPlot}{\code{TRUE} or \code{FALSE} (default), provide an estimate
plot where for each estimator the estimated coefficient and confidence
intervals are plotted.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Moderated mediation
}
