% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIIM3d.R
\name{plotIIM3d}
\alias{plotIIM3d}
\title{Makes plots of Index of moderated mediation of moderatedMediationSem object}
\usage{
plotIIM3d(data, xmmod, mymod, mvars = mvars, res = res)
}
\arguments{
\item{data}{data frame containg the moderators}

\item{xmmod}{moedraotr of x-m path}

\item{mymod}{moderator of m-y path}

\item{mvars}{vector of mediators names}

\item{res}{results of moderatedMediationSem function}
}
\value{
empty, directly plots all indices of mediation
}
\description{
Makes plots of Index of moderated mediation of moderatedMediationSem object
}
