% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemm.R
\name{print.gemm}
\alias{print.gemm}
\title{print method of object of class gemm}
\usage{
\method{print}{gemm}(x, ..., digits = 2)
}
\arguments{
\item{x}{object of class gemm}

\item{...}{additional pars}

\item{digits}{number of digits}
}
\description{
print method of object of class gemm
}
