% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimpleSlopes.R
\name{simpleSlopes}
\alias{simpleSlopes}
\title{Makes simple slop plots of moderatedMediationSem object}
\usage{
simpleSlopes(data, xvar, yvar, mod, mvars, parEst, vdichotomous, modLevels,
  path = NULL)
}
\arguments{
\item{data}{data frame containg the variables of the model}

\item{xvar}{predictor variable name}

\item{yvar}{depedendent variable name}

\item{mod}{moderator name}

\item{mvars}{vector of mediators names}

\item{parEst}{parameter estimates from lavaan results}

\item{vdichotomous}{indicates whether moderator is dichotomous (TRUE)}

\item{modLevels}{levels of dichotomous moderator}

\item{path}{which path is used}
}
\value{
empty, directly plots all simple slopes and all indices of mediation
}
\description{
Makes simple slop plots of moderatedMediationSem object
}
