% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIIM3d.R
\name{IMM3d}
\alias{IMM3d}
\title{Computes Index of moderated mediation of gemm object}
\usage{
IMM3d(M1, M2, parEst = parEst, i = 1)
}
\arguments{
\item{parEst}{parameter estimates from lavaan results}

\item{data}{data frame containg the moderators}

\item{xmmod}{moderator of x-m path}

\item{mymod}{moderator of m-y path}

\item{mvars}{vector of mediators names}
}
\value{
vector of index of moderated mediation with CI limits for a given mediator
}
\description{
Computes Index of moderated mediation of gemm object
}
