% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIIM3d.R
\name{plotIIM3d}
\alias{plotIIM3d}
\title{Makes 3D plots of Index of Moderated Mediation of gemm object}
\usage{
plotIIM3d(x, ...)
}
\arguments{
\item{x}{results of gemm function}

\item{...}{optional}
}
\value{
empty, directly plots all indices of mediation
}
\description{
Makes 3D plots of Index of Moderated Mediation of gemm object
}
